Materials Update and Envy Enemy Prefab
I added folders for each individual sin bosses materials that are used for each, and I have now added in a working envy enemy prefab to the game. I also tampered some of the other sin bosses navMesh Agent to fit to the bottom of the model so the enemy isn't half through the floor.