SlothAI added an some file alterations
The sloth boss now has three phases; the first is the basic melee attack, the second has a speed boost when weak spot is hit, and the third the boss starts throwing their javelin at the player then goes to retrieve it before throwing it again. Messed with files to align with AI's for each boss.