wrathAI and Script Update
I added in code that works the wrath boss where it starts with a melee attack with the sword then after phase two a there is a cooldown that when done the boss will be invincible for a set time, and the third phase the boss starts sprinting. I added comments to help understand all of the code I did and how they function.