Level One Complete
I added in a wave system that spawns two types of enemies for the base waves and then boss after the last wave. Worked on the enemies to actually attack in correct direction. I added a system where you can pick up health packs found on the map of the level. I worked on skybox and UI that was left.